Separate utils into their own files
