--- recipes/utils/micromark-extension-wiki-link.d.ts
@@ -1,3 +1,0 @@
-declare module 'micromark-extension-wiki-link' {
-  function syntax(opts?: unknown): Extension
-}
