+++ recipes/utils/slugify.ts
@@ -1,0 +1,4 @@
+/** Turns a page name into a slug like `my-cool-recipe` */
+export const slugify = (name: string) => {
+  return name.replace(/\s+/g, '-').toLocaleLowerCase()
+}
